feat: 상품 관리 탭의 내용이 다른 탭을 이동하고 돌아와도 그대로 보여지는 기능 구현
상품 관리 탭에서 상품을 추가하고 다른 탭으로 이동후 다시 상품 관리
탭으로 돌아왔을 때, 현재 상품의 내용이 테이블에 표시되는 기능
상품 현황 테이블을 그리는 부분을 외부 모듈로 분리하고 처음 렌더링 할 때
테이블을 그리는 기능 추가
