+++ src/attachProductAddEvent.js
@@ -1,4 +1,5 @@
-import { DOM_ID_SELECTOR, DOM_CLASS_SELECTOR, ERROR_MESSAGE } from './constants.js';
+import { DOM_ID_SELECTOR, ERROR_MESSAGE } from './constants.js';
+import printProductManageTable from './printProductManageTable.js';
 
 const isValidProductName = (name) => {
   if (!name) {
@@ -55,47 +56,6 @@ const clearProductInput = () => {
   document.getElementById(DOM_ID_SELECTOR.productNameInput).value = '';
   document.getElementById(DOM_ID_SELECTOR.productPriceInput).value = '';
   document.getElementById(DOM_ID_SELECTOR.productQuantityInput).value = '';
-};
-
-const makeProductManageTableTd = (className, text) => {
-  const $td = document.createElement('td');
-  $td.className = className;
-  $td.appendChild(document.createTextNode(text));
-
-  return $td;
-};
-
-const makeProductManageItem = (name, price, quantity) => {
-  const $productManageItem = document.createElement('tr');
-  $productManageItem.className = DOM_CLASS_SELECTOR.productManageItem;
-
-  const $productManageName = makeProductManageTableTd(DOM_CLASS_SELECTOR.productManageName, name);
-  const $productManagePrice = makeProductManageTableTd(DOM_CLASS_SELECTOR.productManagePrice, price);
-  const $productManageQuantity = makeProductManageTableTd(DOM_CLASS_SELECTOR.productManageQuantity, quantity);
-
-  $productManageItem.appendChild($productManageName);
-  $productManageItem.appendChild($productManagePrice);
-  $productManageItem.appendChild($productManageQuantity);
-
-  return $productManageItem;
-};
-
-const removeProductManageItems = () => {
-  const $productManageItems = document.getElementsByClassName(DOM_CLASS_SELECTOR.productManageItem);
-
-  while ($productManageItems.length > 0) {
-    $productManageItems[0].remove();
-  }
-};
-
-const printProductManageTable = (productItems) => {
-  const $productManageTable = document.getElementById(DOM_ID_SELECTOR.productManageTable);
-  removeProductManageItems();
-
-  for (let name in productItems) {
-    const $productManageItem = makeProductManageItem(name, productItems[name].price, productItems[name].quantity);
-    $productManageTable.appendChild($productManageItem);
-  }
 };
 
 const attachProductAddEvent = (product) => {
